move builtin_type into builtin.py
