+++ spy/vm/module.py
@@ -2,10 +2,10 @@ from typing import TYPE_CHECKING, Optional, Iterable
 from spy.fqn import QN, FQN
 from spy.vm.primitive import W_Void
 from spy.vm.b import B
-from spy.vm.object import W_Object, builtin_type, W_Type, W_Dynamic
+from spy.vm.object import W_Object, W_Type, W_Dynamic
 from spy.vm.str import W_Str
 from spy.vm.function import W_ASTFunc
-from spy.vm.builtin import builtin_func
+from spy.vm.builtin import builtin_func, builtin_type
 from spy.vm.opimpl import W_OpImpl, W_OpArg
 
 if TYPE_CHECKING:
