+++ spy/vm/modules/unsafe/__init__.py
@@ -2,7 +2,7 @@ from typing import TYPE_CHECKING
 import fixedint
 from spy.vm.primitive import W_I32, W_Void
 from spy.vm.b import B
-from spy.vm.object import builtin_type
+from spy.vm.builtin import builtin_type
 from spy.vm.w import W_Object, W_Type
 from spy.vm.registry import ModuleRegistry
 from spy.vm.opimpl import W_OpImpl, W_OpArg
