+++ spy/vm/object.py
@@ -44,7 +44,6 @@ For simple cases, SPy app-level types are instances of W_Type, which is
 basically a thin wrapper around the correspindig interp-level W_* class.
 """
 
-import typing
 from typing import (TYPE_CHECKING, ClassVar, Type, Any, Annotated, Optional,
                     Union)
 from spy.fqn import QN
@@ -392,25 +391,3 @@ def synthesize_meta_op_CALL(pyclass: Type[W_Object]) -> Any:
 
     return meta_op_CALL
 
-def builtin_type(name: str) -> Any:
-    """
-    Class decorator to simplify the creation of SPy types.
-
-    Given a W_* class, it automatically creates the corresponding instance of
-    W_Type and attaches it to the W_* class.
-    """
-    def decorator(pyclass: Type[W_Object]) -> Type[W_Object]:
-        W_MetaClass = make_metaclass(name, pyclass)
-
-        pyclass._w = W_MetaClass(name, pyclass)
-        # setup __spy_members__
-        pyclass.__spy_members__ = {}
-        for field, t in pyclass.__annotations__.items():
-            member = _get_member_maybe(t)
-            if member is not None:
-                member.field = field
-                member.w_type = typing.get_args(t)[0]._w
-                pyclass.__spy_members__[member.name] = member
-
-        return pyclass
-    return decorator
